+++ changelog/changelog.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 2026-3-?? (? Version, 0.7)
+### Improved
+* Proposition 13
+	* eccentricity control integrated into the model with point A
+
 ## 2026-2-26 (Olive Version, 0.6)
 ### Improved
 * Lemma 11
@@ -40,7 +45,7 @@
 	* minor bug fixes
 	
 * Propositions 10, 11
-	* eccentricity of both models controlled by point A
+	* eccentricity control integrated into the model with point A
 
 ## 2025-12-17 (Green Version, 0.3)
 ### Improved